Atualizado os serializers da JsonConfiguration e review no pom.xml
