+++ src/main/java/discord/gdd/config/json/serializers/LocationSerializer.java
@@ -1,48 +1,27 @@
 package discord.gdd.config.json.serializers;
 
-import com.google.gson.*;
-import discord.gdd.config.json.exceptions.InvalidWorldException;
-import org.bukkit.Bukkit;
+import com.google.gson.JsonElement;
+import com.google.gson.JsonObject;
+import com.google.gson.JsonSerializationContext;
+import com.google.gson.JsonSerializer;
 import org.bukkit.Location;
-import org.bukkit.World;
 
 import java.lang.reflect.Type;
-import java.util.HashMap;
-import java.util.Map;
 
-public class LocationSerializer implements JsonSerializer<Location>, JsonDeserializer<Location> {
+public class LocationSerializer implements JsonSerializer<Location> {
     @Override
     public JsonElement serialize(Location src, Type typeOfSrc, JsonSerializationContext context) {
-        Map<String, String> loc = new HashMap<>();
-        loc.put("world", src.getWorld().getName());
-        loc.put("x", src.getX() + "");
-        loc.put("y", src.getY() + "");
-        loc.put("z", src.getZ() + "");
-        loc.put("yaw", src.getYaw() + "");
-        loc.put("pitch", src.getPitch() + "");
-        return context.serialize(loc);
-    }
+        JsonObject locJson = new JsonObject();
 
-    @Override
-    public Location deserialize(JsonElement json, Type typeOfT, JsonDeserializationContext context) throws JsonParseException {
-        JsonObject jObj = json.getAsJsonObject();
-        if (!jObj.has("world")
-                || !jObj.has("x") || !jObj.has("y") || !jObj.has("z")
-                || !jObj.has("yaw") || !jObj.has("pitch"))
-            throw new JsonParseException("Localização invalida!");
+        locJson.addProperty("world", src.getWorld().getName());
+        locJson.addProperty("x", src.getX() + "");
+        locJson.addProperty("y", src.getY() + "");
+        locJson.addProperty("z", src.getZ() + "");
+        if (src.getYaw() != 0.0)
+            locJson.addProperty("yaw", src.getYaw() + "");
+        if (src.getPitch() != 0.0)
+            locJson.addProperty("pitch", src.getPitch() + "");
 
-        return new Location(getWorld(jObj.get("world").getAsString()),
-                jObj.get("x").getAsDouble(),
-                jObj.get("y").getAsDouble(),
-                jObj.get("z").getAsDouble(),
-                jObj.get("yaw").getAsFloat(),
-                jObj.get("pitch").getAsFloat()
-                );
-    }
-
-    private World getWorld(String nome) throws InvalidWorldException {
-        World world = Bukkit.getWorld(nome);
-        if (world == null) throw new InvalidWorldException(nome);
-        return world;
+        return locJson;
     }
 }
